% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/803.Association-measures-OR.R
\name{Association.Measures.OR}
\alias{Association.Measures.OR}
\title{This function  implements 10 methods for OR confidence interaval - Woolf, Gart, Agresti,
 MOVER, MI_NU, CornEX, CornMP, BaPiEx, BaPiMP, Inverse Sinh and Inverse Sinh Adjusted}
\usage{
Association.Measures.OR(ai, bi, ci, di, alp, e1, e2)
}
\arguments{
\item{ai}{- Numeric value representing True Positives}

\item{bi}{- Numeric value representing False Positives}

\item{ci}{- Numeric value representing False Negatives}

\item{di}{- Numeric value representing True Negatives}

\item{alp}{- Numeric value representing required significance level}

\item{e1}{- Numeric value representing adjustment factor for inverse Sinh adjusted}

\item{e2}{- Numeric value representing adjustment factor for inverse Sinh adjusted}
}
\value{
A dataframe with
 \item{Method}{Ten methods are implemented - Woolf, Gart, Agresti, MOVER, MI_NU, CornEX,
 CornMP, BaPiEx, BaPiMP, Inverse Sinh, Inverse Sinh Adjusted}
 \item{Estimate}{The estimated value}
 \item{Lower.Limit}{The lower limit }
 \item{Upper.Limit}{The upper limit }
}
\description{
This function  implements 10 methods for OR confidence interaval - Woolf, Gart, Agresti,
 MOVER, MI_NU, CornEX, CornMP, BaPiEx, BaPiMP, Inverse Sinh and Inverse Sinh Adjusted
}
\details{
Ten methods - Woolf, Gart, Agresti, MOVER, MI_NU, CornEX, CornMP,
BaPiEx, BaPiMP, Inverse Sinh, Inverse Sinh Adjusted are implimented from
ORCI package (9 methods) and PropCIs package (1 method-Miettinen-Nurminen).
}
\examples{
ai=7; bi=11; ci=1; di=17; alp=0.05; e1=0.5; e2=0.5
Association.Measures.OR(ai, bi, ci, di, alp, e1, e2)
}
\references{
[1] Fagerland, M. W., & Newcombe, R. G. (2013).
Confidence intervals for odds ratio and relative risk based on the inverse hyperbolic sine transformation.
Statistics in medicine, 32(16), 2823-2836.
[2] Upton, G. J. (2016).
Categorical Data Analysis by Example.
John Wiley & Sons.
}
\seealso{
\code{\link[ORCI]{Woolf.CI}{Gart.CI}{Agrestiind.CI}{MOVER.CI}{Cornfieldexact.CI}{Cornfieldmidp.CI}{BPexact.CI}{BPmidp.CI}{Invsinh.CI}} with additional parameters. This function calls ORCI functions.
   \code{\link[PropCIs]{orscoreci}} with additional parameters. This function calls orscoreci function.

Other Association Measures methods: \code{\link{Association.Measures.RD}},
  \code{\link{Association.Measures.RR}}
}
