% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/801.Association-measures-RD.R
\name{Association.Measures.RD}
\alias{Association.Measures.RD}
\title{This function  calls pairwiseCI package and 3 methods from that are implemented with consistant output view.}
\usage{
Association.Measures.RD(ai, bi, ci, di, alp)
}
\arguments{
\item{ai}{- Numeric value representing True Positives}

\item{bi}{- Numeric value representing False Positives}

\item{ci}{- Numeric value representing False Negatives}

\item{di}{- Numeric value representing True Negatives}

\item{alp}{- Numeric value representing required significance level}
}
\value{
A dataframe with
 \item{Method}{Three methods are implemented - Newcombes Hybrid score, Wald continuity corrected and Agresti Caffo}
 \item{Estimate}{The estimated value}
 \item{Lower.Limit}{The lower limit }
 \item{Upper.Limit}{The upper limit }
}
\description{
This function  calls pairwiseCI package and 3 methods from that are implemented with consistant output view.
}
\details{
Three methods from pairwiseCI package is implemented with consistant output view.
Methods implemented are Newcombes Hybrid score, Wald continuity corrected and Agresti Caffo.
}
\examples{
ai=7; bi=11; ci=1; di=17; alp=0.05 # Example from [reference 1]
suppressWarnings(Association.Measures.RD(ai,bi,ci,di,alp))
}
\references{
[1] Fagerland, M. W., & Newcombe, R. G. (2013).
Confidence intervals for odds ratio and relative risk based on the inverse hyperbolic sine transformation.
Statistics in medicine, 32(16), 2823-2836.
[2] Upton, G. J. (2016).
Categorical Data Analysis by Example.
John Wiley & Sons.
}
\seealso{
\code{\link[pairwiseCI]{Prop.diff}}  provides this functionality with additional parameters as well

Other Association Measures methods: \code{\link{Association.Measures.OR}},
  \code{\link{Association.Measures.RR}}
}
