% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/802.Association-measures-RR.R
\name{Association.Measures.RR}
\alias{Association.Measures.RR}
\title{This function  calculates Risk Ratio and has 6 methods implemented leveraging pairwiseCI and metafor packags;
 Further inverse sinh base and adjusted are made available.}
\usage{
Association.Measures.RR(ai, bi, ci, di, alp, c1, c2, c3, c4)
}
\arguments{
\item{ai}{- Numeric value representing True Positives}

\item{bi}{- Numeric value representing False Positives}

\item{ci}{- Numeric value representing False Negatives}

\item{di}{- Numeric value representing True Negatives}

\item{alp}{- Numeric value representing required significance level}

\item{c1}{- Numeric value representing adjustment factor for inverse Sinh adjusted}

\item{c2}{- Numeric value representing adjustment factor for inverse Sinh adjusted}

\item{c3}{- Numeric value representing adjustment factor for inverse Sinh adjusted}

\item{c4}{- Numeric value representing adjustment factor for inverse Sinh adjusted}
}
\value{
A dataframe with
 \item{Method}{Six methods are implemented - Woolf, Gart, MI_NU, MOVER, ISH and ISH.Adjusted}
 \item{Estimate}{The estimated value}
 \item{Lower.Limit}{The lower limit }
 \item{Upper.Limit}{The upper limit }
}
\description{
This function  calculates Risk Ratio and has 6 methods implemented leveraging pairwiseCI and metafor packags;
 Further inverse sinh base and adjusted are made available.
}
\details{
Two methods from pairwiseCI package and two methods from metafor package is implemented
along with inverese Sinh and inverse Sinh adjusted with consistant output view.
Methods implemented are Woolf, Gart, MI_NU, MOVER, Inverse Sinh and Inverse Sinh Adjusted.
}
\examples{
## Example taken from Fagerland and Newcombe [reference 1]
ai=7; bi=11; ci=1; di=17; alp=0.05
c1=0; c2=0.2; c3=0; c4=0.8
suppressWarnings(Association.Measures.RR(ai,bi,ci,di,alp, c1, c2, c3, c4))
}
\references{
[1] Fagerland, M. W., & Newcombe, R. G. (2013).
Confidence intervals for odds ratio and relative risk based on the inverse hyperbolic sine transformation.
Statistics in medicine, 32(16), 2823-2836.
[2] Upton, G. J. (2016).
Categorical Data Analysis by Example.
John Wiley & Sons.
}
\seealso{
\code{\link[pairwiseCI]{Prop.ratio}}  provides Score, MNScore, MOVER and GNC. This function call pairwiseCI functions.
   \code{\link[metafor]{escalc}}  provides comprenhisive functionality from which RR alone is used in this function

Other Association Measures methods: \code{\link{Association.Measures.OR}},
  \code{\link{Association.Measures.RD}}
}
