% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/304.Internal-methods.R
\name{Calculate.cc.OR}
\alias{Calculate.cc.OR}
\title{This Function to calculate OR and the point of change}
\usage{
Calculate.cc.OR(cc.df, cc.value)
}
\arguments{
\item{cc.df}{- description}

\item{cc.value}{- description}
}
\value{
A dataframe with
 \item{r.df}{Sequence}
}
\description{
This Function to calculate OR and the point of change
}
\seealso{
Other Internal methods: \code{\link{Calculate.cc.RD}},
  \code{\link{Calculate.cc.RR}},
  \code{\link{Find.RS.metric}},
  \code{\link{Find.rma.metric}},
  \code{\link{Find.sequence.rma}},
  \code{\link{Find.sequence}}, \code{\link{GSquared}},
  \code{\link{In.Each.Row}}, \code{\link{gmat}},
  \code{\link{is.between}},
  \code{\link{pass2.In.Each.Row.new.cc.or}},
  \code{\link{pass2.In.Each.Row.new.cc.rd}},
  \code{\link{pass2.In.Each.Row.new.cc.rr}}
}
\keyword{internal}
