% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/402.Bayesian-EstimationMethod.R
\name{ci.BMDU}
\alias{ci.BMDU}
\title{Bayes estimation with different hyper priors}
\usage{
ci.BMDU(x, d)
}
\arguments{
\item{x}{- Vector of positive integers}

\item{d}{- Number of divisions}
}
\value{
A list of dataframes
 \item{BE.diff.prior.vol.df}{ Bayesian estimation Volume}
 \item{BE.diff.prior.LL.df}{  Dataframe of Bayesian prior Lower Limits}
 \item{BE.diff.prior.UL.df}{  Dataframe of Bayesian prior Upper Limits}
 \item{BE.diff.prior.mean.df}{  Dataframe of Bayesian prior mean}
}
\description{
Bayes estimation with different hyper priors
}
\examples{
x = c(56,72,73,59,62,87,68,99,98)
d=3
ci.BMDU(x,d)
}
\seealso{
\code{\link[rBeta2009]{rdirichlet} with additional n and shape paramaters. This function calls rBeta2009 function.}

Other Confidence Interval for Multinomial Proportion: \code{\link{ci.FS}},
  \code{\link{ci.GM}}, \code{\link{ci.QH}},
  \code{\link{ci.SG}}, \code{\link{ci.WS}},
  \code{\link{ci.WaldCC}}, \code{\link{ci.Wald}}
}
