% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/400.ConsolidatedEstimationMethods.R
\name{ci.FS}
\alias{ci.FS}
\title{The simultaneous confidence interval for multinomial proportions based on the
method proposed in Fitzpatrick and Scott (1987)}
\usage{
ci.FS(inpmat, alpha)
}
\arguments{
\item{inpmat}{- The input matrix}

\item{alpha}{- Alpha value (significance level required)}
}
\value{
A list of dataframes
 \item{WaldCC.Volume}{ Wald Volume}
 \item{WaldCC.UpLim}{  Dataframe of Wald Upper Limits}
 \item{WaldCC.LowLim}{  Dataframe of Wald Lower Limits}
 \item{WaldCC.Length}{  Dataframe of Wald Lengths}
}
\description{
The simultaneous confidence interval for multinomial proportions based on the
method proposed in Fitzpatrick and Scott (1987)
}
\examples{
x = c(56,72,73,59,62,87,68,99,98)
inpmat = cbind(x[1:3],x[4:6],x[7:9])
alpha=0.05
ci.FS(inpmat,alpha)
}
\references{
[1] Fitzpatrick, S. and Scott, A. (1987).
Quick simultaneous confidence interval for multinomial proportions.
Journal of American Statistical Association 82(399): 875-878.
}
\seealso{
Other Confidence Interval for Multinomial Proportion: \code{\link{ci.BMDU}},
  \code{\link{ci.GM}}, \code{\link{ci.QH}},
  \code{\link{ci.SG}}, \code{\link{ci.WS}},
  \code{\link{ci.WaldCC}}, \code{\link{ci.Wald}}
}
\author{
Subbiah and Balakrishna S Kesavan
}
