% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/400.ConsolidatedEstimationMethods.R
\name{ci.GM}
\alias{ci.GM}
\title{The simultaneous confidence interval for multinomial proportions based on the method proposed in Goodman (1965)}
\usage{
ci.GM(inpmat, alpha)
}
\arguments{
\item{inpmat}{- The input matrix}

\item{alpha}{- Alpha value (significance level required)}
}
\value{
A list of dataframes
 \item{GM.Volume}{ GM Volume}
 \item{GM.UpLim}{  Dataframe of GM Upper Limits}
 \item{GM.LowLim}{  Dataframe of GM Lower Limits}
 \item{GM.Length}{  Dataframe of GM Lengths}
}
\description{
The simultaneous confidence interval for multinomial proportions based on the method proposed in Goodman (1965)
}
\examples{
x = c(56,72,73,59,62,87,68,99,98)
inpmat = cbind(x[1:3],x[4:6],x[7:9])
alpha=0.05
ci.GM(inpmat,alpha)
}
\references{
[1] Goodman, L.A. (1965).
On Simultaneous Confidence Intervals for Multinomial Proportions.
Technometrics 7: 247-254.
}
\seealso{
Other Confidence Interval for Multinomial Proportion: \code{\link{ci.BMDU}},
  \code{\link{ci.FS}}, \code{\link{ci.QH}},
  \code{\link{ci.SG}}, \code{\link{ci.WS}},
  \code{\link{ci.WaldCC}}, \code{\link{ci.Wald}}
}
\author{
Subbiah and Balakrishna S Kesavan
}
