% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/401.SG-EstimationMethod.R
\name{ci.SG}
\alias{ci.SG}
\title{The simultaneous confidence interval for multinomial proportions based on the method
proposed in Sison and Glaz (1995)}
\usage{
ci.SG(inpmat, alpha)
}
\arguments{
\item{inpmat}{- The input matrix}

\item{alpha}{- Alpha value (significance level required)}
}
\value{
A list of dataframes
 \item{SG.Volume}{ SG Volume}
 \item{SG.Orig.LowLim}{  Dataframe of SG Original Lower Limits}
 \item{SG.Orig.UpLim}{  Dataframe of SG Original Upper Limits}
 \item{SG.Adj.LowLim}{  Dataframe of SG Adjusted Lower Limits}
 \item{SG.Adj.UpLim}{  Dataframe of SG Adjusted Upper Limits}
}
\description{
The simultaneous confidence interval for multinomial proportions based on the method
proposed in Sison and Glaz (1995)
}
\examples{
x = c(56,72,73,59,62,87,68,99,98)
inpmat = cbind(x[1:3],x[4:6],x[7:9])
alpha=0.05
ci.SG(inpmat,alpha)
}
\seealso{
Other Confidence Interval for Multinomial Proportion: \code{\link{ci.BMDU}},
  \code{\link{ci.FS}}, \code{\link{ci.GM}},
  \code{\link{ci.QH}}, \code{\link{ci.WS}},
  \code{\link{ci.WaldCC}}, \code{\link{ci.Wald}}
}
\author{
Subbiah and Balakrishna S Kesavan
}
