% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/400.ConsolidatedEstimationMethods.R
\name{ci.WaldCC}
\alias{ci.WaldCC}
\title{The simple Wald type interval with continuity corrections for multinomial proportions which
is symmetrical about the sample proportions.}
\usage{
ci.WaldCC(inpmat, alpha)
}
\arguments{
\item{inpmat}{- The input matrix}

\item{alpha}{- Alpha value (significance level required)}
}
\value{
A list of dataframes
 \item{WaldCC.Volume}{ Wald Volume}
 \item{WaldCC.UpLim}{  Dataframe of Wald Upper Limits}
 \item{WaldCC.LowLim}{  Dataframe of Wald Lower Limits}
 \item{WaldCC.Length}{  Dataframe of Wald Lengths}
}
\description{
The simple Wald type interval with continuity corrections for multinomial proportions which
is symmetrical about the sample proportions.
}
\examples{
x = c(56,72,73,59,62,87,68,99,98)
inpmat = cbind(x[1:3],x[4:6],x[7:9])
alpha=0.05
ci.WaldCC(inpmat,alpha)
}
\seealso{
Other Confidence Interval for Multinomial Proportion: \code{\link{ci.BMDU}},
  \code{\link{ci.FS}}, \code{\link{ci.GM}},
  \code{\link{ci.QH}}, \code{\link{ci.SG}},
  \code{\link{ci.WS}}, \code{\link{ci.Wald}}
}
\author{
Subbiah and Balakrishna S Kesavan
}
