% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/400.ConsolidatedEstimationMethods.R
\name{ci.WS}
\alias{ci.WS}
\title{The simultaneous confidence interval for multinomial proportions based on the method
proposed in Wilson (1927)}
\usage{
ci.WS(inpmat, alpha)
}
\arguments{
\item{inpmat}{- The input matrix}

\item{alpha}{- Alpha value (significance level required)}
}
\value{
A list of dataframes
 \item{WaldCC.Volume}{ Wald Volume}
 \item{WaldCC.UpLim}{  Dataframe of Wald Upper Limits}
 \item{WaldCC.LowLim}{  Dataframe of Wald Lower Limits}
 \item{WaldCC.Length}{  Dataframe of Wald Lengths}
}
\description{
The simultaneous confidence interval for multinomial proportions based on the method
proposed in Wilson (1927)
}
\examples{
x = c(56,72,73,59,62,87,68,99,98)
inpmat = cbind(x[1:3],x[4:6],x[7:9])
alpha=0.05
ci.WS(inpmat,alpha)
}
\references{
[1] E.B. Wilson,
Probable inference, the law of succession and statistical inference,
J.Am. Stat. Assoc.22 (1927) 209-212.
}
\seealso{
Other Confidence Interval for Multinomial Proportion: \code{\link{ci.FS}},
  \code{\link{ci.GM}}, \code{\link{ci.QH}},
  \code{\link{ci.WaldCC}}, \code{\link{ci.Wald}}
}
