% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/701.2x2-tests.R
\name{exact2x2tests.hypergeom}
\alias{exact2x2tests.hypergeom}
\title{This function  implements 9 methods for testing 2x2 tables and outputs the p-value
 In hypergemetric notation out of N = m+n items r size is chosen; x is observed success}
\usage{
exact2x2tests.hypergeom(x, m, n, r)
}
\arguments{
\item{x}{- Numeric value representing no of success in attribute 1.}

\item{m}{- Numeric value representing Total no of attribute 1}

\item{n}{- Numeric value representing Total no of attribute 2}

\item{r}{- Numeric value representing Total number of success}
}
\value{
A dataframe with
 \item{IRWIN}{Fisher exact test with 'minlike'}
 \item{Central}{Fisher exact test with 'central'}
 \item{Blaker.exact}{Blaker exact test }
 \item{MP_IR}{Fisher exact test with 'minlike', hypergeometric}
 \item{MP_BK}{Blaker exact test, hypergeometric}
 \item{Exact.less}{Fisher exact test if one sided is less}
 \item{Exact.greater}{Fisher exact test if one sided is greater}
 \item{Pearson.Chisq}{Pearson Chisquared test}
 \item{Pearson.Chisq.Yates.cc}{Pearson Chisquared test with Yates continuity correction}
}
\description{
This function  implements 9 methods for testing 2x2 tables and outputs the p-value
 In hypergemetric notation out of N = m+n items r size is chosen; x is observed success
}
\details{
Nine methods - IRWIN, Central, Blaker.exact, MP_IR, MP_BK, Exact.less,
Exact.greater, Pearson.Chisq, Pearson.Chisq.Yates.cc are implimented from
exact2x2 package (7 methods) and 2 methods from base stats
}
\examples{
## Example taken from -p28 Ex3.1 of Upton, G. J. (2016).
## Values are x,m,n,r: related to a hypergeometric distribution (?dhyper, for further help)
x=1; m=3; n=10; r=4
exact2x2tests.hypergeom(x,m,n,r)
}
\references{
[1] Upton, G. J. (2016).
Categorical Data Analysis by Example.
John Wiley & Sons.
}
\seealso{
\code{\link[exact2x2]{fisher.exact}{blaker.exact}} with additional parameters. This function calls exact2x2 functions.
   \code{\link[stats]{chisq.test}} with additional parameters. This function calls base stats function.

Other Test methods: \code{\link{Partition.table}},
  \code{\link{Proportion.diagnostic.test}},
  \code{\link{exact2x2tests.regular}}
}
