% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/701.2x2-tests.R
\name{exact2x2tests.regular}
\alias{exact2x2tests.regular}
\title{This function  implements 9 methods for testing 2x2 tables and outputs the p-value
 This is the regular function taking the 2x2 table as input}
\usage{
exact2x2tests.regular(ai, bi, ci, di)
}
\arguments{
\item{ai}{- Numeric value representing True Positives}

\item{bi}{- Numeric value representing False Positives}

\item{ci}{- Numeric value representing False Negatives}

\item{di}{- Numeric value representing True Negatives}
}
\value{
A dataframe with
 \item{IRWIN}{Fisher exact test with 'minlike'}
 \item{Central}{Fisher exact test with 'central'}
 \item{Blaker.exact}{Blaker exact test }
 \item{MP_IR}{Fisher exact test with 'minlike', hypergeometric}
 \item{MP_BK}{Blaker exact test, hypergeometric}
 \item{Exact.less}{Fisher exact test if one sided is less}
 \item{Exact.greater}{Fisher exact test if one sided is greater}
 \item{Pearson.Chisq}{Pearson Chisquared test}
 \item{Pearson.Chisq.Yates.cc}{Pearson Chisquared test with Yates continuity correction}
}
\description{
This function  implements 9 methods for testing 2x2 tables and outputs the p-value
 This is the regular function taking the 2x2 table as input
}
\details{
Nine methods - IRWIN, Central, Blaker.exact, MP_IR, MP_BK, Exact.less,
Exact.greater, Pearson.Chisq, Pearson.Chisq.Yates.cc are implimented from
exact2x2 package (7 methods) and 2 methods from base stats
}
\examples{
## Example taken from -p28 Ex3.1 of Upton, G. J. (2016).
ai=1; bi=3; ci=10; di=4
exact2x2tests.regular(ai,bi,ci,di)
}
\references{
[1] Upton, G. J. (2016).
Categorical Data Analysis by Example.
John Wiley & Sons.
}
\seealso{
\code{\link[exact2x2]{fisher.exact}{blaker.exact}} with additional parameters. This function calls exact2x2 functions.
   \code{\link[stats]{chisq.test}} with additional parameters. This function calls base stats function.

Other Test methods: \code{\link{Partition.table}},
  \code{\link{Proportion.diagnostic.test}},
  \code{\link{exact2x2tests.hypergeom}}
}
