% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/304.Internal-methods.R
\name{Find.rma.metric}
\alias{Find.rma.metric}
\title{This function finds the rma OR/RR/RD for the given  range using metafor}
\usage{
Find.rma.metric(a, b, c, d, RS.seq, metric)
}
\arguments{
\item{a}{- description}

\item{b}{- description}

\item{c}{- description}

\item{d}{- description}

\item{RS.seq}{- description}

\item{metric}{- description}
}
\value{
A dataframe with
 \item{output}{Sequence}
}
\description{
This function finds the rma OR/RR/RD for the given  range using metafor
}
\seealso{
Other Internal methods: \code{\link{Calculate.cc.OR}},
  \code{\link{Calculate.cc.RD}},
  \code{\link{Calculate.cc.RR}},
  \code{\link{Find.RS.metric}},
  \code{\link{Find.sequence.rma}},
  \code{\link{Find.sequence}}, \code{\link{GSquared}},
  \code{\link{In.Each.Row}}, \code{\link{gmat}},
  \code{\link{is.between}},
  \code{\link{pass2.In.Each.Row.new.cc.or}},
  \code{\link{pass2.In.Each.Row.new.cc.rd}},
  \code{\link{pass2.In.Each.Row.new.cc.rr}}
}
\keyword{internal}
