% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/301.Find.RS.metric.R
\name{Find.RS.metric}
\alias{Find.RS.metric}
\title{Find the point of reversal of OR/RR/RD}
\usage{
Find.RS.metric(Figure.df)
}
\arguments{
\item{Figure.df}{- description}
}
\value{
A dataframe with
 \item{x1}{Sequence}
}
\description{
Find the point of reversal of OR/RR/RD
}
\seealso{
Other Internal methods: \code{\link{Calculate.cc.OR}},
  \code{\link{Calculate.cc.RD}},
  \code{\link{Calculate.cc.RR}},
  \code{\link{Find.rma.metric}},
  \code{\link{Find.sequence.rma}},
  \code{\link{Find.sequence}}, \code{\link{GSquared}},
  \code{\link{In.Each.Row}}, \code{\link{gmat}},
  \code{\link{is.between}},
  \code{\link{pass2.In.Each.Row.new.cc.or}},
  \code{\link{pass2.In.Each.Row.new.cc.rd}},
  \code{\link{pass2.In.Each.Row.new.cc.rr}}
}
\keyword{internal}
