% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/301.Find.RS.metric.R
\name{Find.sequence}
\alias{Find.sequence}
\title{Find the sequence for the given  range}
\usage{
Find.sequence(RHS.value, LHS.value)
}
\arguments{
\item{LHS.value}{- description}

\item{RHS/value}{- description}
}
\value{
A dataframe with
 \item{x1}{Sequence}
}
\description{
Find the sequence for the given  range
}
\seealso{
Other Internal methods: \code{\link{Calculate.cc.OR}},
  \code{\link{Calculate.cc.RD}},
  \code{\link{Calculate.cc.RR}},
  \code{\link{Find.RS.metric}},
  \code{\link{Find.rma.metric}},
  \code{\link{Find.sequence.rma}}, \code{\link{GSquared}},
  \code{\link{In.Each.Row}}, \code{\link{gmat}},
  \code{\link{is.between}},
  \code{\link{pass2.In.Each.Row.new.cc.or}},
  \code{\link{pass2.In.Each.Row.new.cc.rd}},
  \code{\link{pass2.In.Each.Row.new.cc.rr}}
}
\keyword{internal}
