% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/304.Internal-methods.R
\name{Find.sequence.rma}
\alias{Find.sequence.rma}
\title{This function  finds the sequence for graph}
\usage{
Find.sequence.rma(RS.val)
}
\arguments{
\item{RS.seq}{- Sequence of values for cc}
}
\value{
A dataframe with
 \item{Ret.seq}{Sequence}
}
\description{
This function  finds the sequence for graph
}
\seealso{
Other Internal methods: \code{\link{Calculate.cc.OR}},
  \code{\link{Calculate.cc.RD}},
  \code{\link{Calculate.cc.RR}},
  \code{\link{Find.RS.metric}},
  \code{\link{Find.rma.metric}},
  \code{\link{Find.sequence}}, \code{\link{GSquared}},
  \code{\link{In.Each.Row}}, \code{\link{gmat}},
  \code{\link{is.between}},
  \code{\link{pass2.In.Each.Row.new.cc.or}},
  \code{\link{pass2.In.Each.Row.new.cc.rd}},
  \code{\link{pass2.In.Each.Row.new.cc.rr}}
}
\keyword{internal}
