% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/704.Internal-functions.R
\name{gmat}
\alias{gmat}
\title{This partition calcluation}
\usage{
gmat(a, b, ie, je, d)
}
\arguments{
\item{a}{- description}

\item{b}{- description}

\item{ie}{- description}

\item{je}{- description}

\item{d}{- description}
}
\value{
A dataframe with
 \item{x1}{Sequence}
}
\description{
This partition calcluation
}
\seealso{
Other Internal methods: \code{\link{Calculate.cc.OR}},
  \code{\link{Calculate.cc.RD}},
  \code{\link{Calculate.cc.RR}},
  \code{\link{Find.RS.metric}},
  \code{\link{Find.rma.metric}},
  \code{\link{Find.sequence.rma}},
  \code{\link{Find.sequence}}, \code{\link{GSquared}},
  \code{\link{In.Each.Row}}, \code{\link{is.between}},
  \code{\link{pass2.In.Each.Row.new.cc.or}},
  \code{\link{pass2.In.Each.Row.new.cc.rd}},
  \code{\link{pass2.In.Each.Row.new.cc.rr}}
}
\keyword{internal}
