% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/704.Internal-functions.R
\name{GSquared}
\alias{GSquared}
\title{This G squared metric}
\usage{
GSquared(catd)
}
\arguments{
\item{catd}{- description}
}
\value{
A list with
 \item{x1}{Sequence}
}
\description{
This G squared metric
}
\seealso{
Other Internal methods: \code{\link{Calculate.cc.OR}},
  \code{\link{Calculate.cc.RD}},
  \code{\link{Calculate.cc.RR}},
  \code{\link{Find.RS.metric}},
  \code{\link{Find.rma.metric}},
  \code{\link{Find.sequence.rma}},
  \code{\link{Find.sequence}}, \code{\link{In.Each.Row}},
  \code{\link{gmat}}, \code{\link{is.between}},
  \code{\link{pass2.In.Each.Row.new.cc.or}},
  \code{\link{pass2.In.Each.Row.new.cc.rd}},
  \code{\link{pass2.In.Each.Row.new.cc.rr}}
}
\keyword{internal}
