% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/304.Internal-methods.R
\name{In.Each.Row}
\alias{In.Each.Row}
\title{This Function to calculate OR/RR/RD and the LHS/RHS of change}
\usage{
In.Each.Row(odf.row, cc.vals)
}
\arguments{
\item{odf.row}{- description}

\item{cc.vals}{- description}
}
\value{
A dataframe with
 \item{Full.Slevel.df}{Sequence}
}
\description{
This Function to calculate OR/RR/RD and the LHS/RHS of change
}
\seealso{
Other Internal methods: \code{\link{Calculate.cc.OR}},
  \code{\link{Calculate.cc.RD}},
  \code{\link{Calculate.cc.RR}},
  \code{\link{Find.RS.metric}},
  \code{\link{Find.rma.metric}},
  \code{\link{Find.sequence.rma}},
  \code{\link{Find.sequence}}, \code{\link{GSquared}},
  \code{\link{gmat}}, \code{\link{is.between}},
  \code{\link{pass2.In.Each.Row.new.cc.or}},
  \code{\link{pass2.In.Each.Row.new.cc.rd}},
  \code{\link{pass2.In.Each.Row.new.cc.rr}}
}
\keyword{internal}
