% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/304.Internal-methods.R
\name{is.between}
\alias{is.between}
\title{This Is between utility function}
\usage{
is.between(x, a, b)
}
\arguments{
\item{x}{- description}

\item{a}{- description}

\item{b}{- description}
}
\value{
A dataframe with
 \item{r.df}{Sequence}
}
\description{
This Is between utility function
}
\seealso{
Other Internal methods: \code{\link{Calculate.cc.OR}},
  \code{\link{Calculate.cc.RD}},
  \code{\link{Calculate.cc.RR}},
  \code{\link{Find.RS.metric}},
  \code{\link{Find.rma.metric}},
  \code{\link{Find.sequence.rma}},
  \code{\link{Find.sequence}}, \code{\link{GSquared}},
  \code{\link{In.Each.Row}}, \code{\link{gmat}},
  \code{\link{pass2.In.Each.Row.new.cc.or}},
  \code{\link{pass2.In.Each.Row.new.cc.rd}},
  \code{\link{pass2.In.Each.Row.new.cc.rr}}
}
\keyword{internal}
