% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/301.Find.RS.metric.R
\name{pass2.In.Each.Row.new.cc.or}
\alias{pass2.In.Each.Row.new.cc.or}
\title{Find the OR,RR and RD for each row with new CC range}
\usage{
pass2.In.Each.Row.new.cc.or(odf.row, cc.vals)
}
\arguments{
\item{odf.row}{- description}

\item{cc.vals}{- description}
}
\value{
A dataframe with
 \item{x1}{Sequence}
}
\description{
Find the OR,RR and RD for each row with new CC range
}
\seealso{
Other Internal methods: \code{\link{Calculate.cc.OR}},
  \code{\link{Calculate.cc.RD}},
  \code{\link{Calculate.cc.RR}},
  \code{\link{Find.RS.metric}},
  \code{\link{Find.rma.metric}},
  \code{\link{Find.sequence.rma}},
  \code{\link{Find.sequence}}, \code{\link{GSquared}},
  \code{\link{In.Each.Row}}, \code{\link{gmat}},
  \code{\link{is.between}},
  \code{\link{pass2.In.Each.Row.new.cc.rd}},
  \code{\link{pass2.In.Each.Row.new.cc.rr}}
}
\keyword{internal}
