% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/303.Generate-graph-functions.R
\name{Plot.reversals.RR}
\alias{Plot.reversals.RR}
\title{############################################################################################################
 Given an 2x2 table with ai, bi, ci and di, if there are any zero cells, a continuity correction
 (CC) is added upto 0.5. This function checks if this addition of continuity correction causes
 the reversing the Odds Ratio, Relative Risk and Risk Difference.
 We provide 2 ways to add this CC. Adding CC to all the cells or only to the zero cells.
 The function will identify if there is a reversal and also give the point of reversal.
 The continuity correction can be also graphed if the option is selected.}
\usage{
Plot.reversals.RR(ai, bi, ci, di, With.CI = FALSE)
}
\arguments{
\item{ai}{- Numeric value}

\item{bi}{- Numeric value}

\item{ci}{- Numeric value}

\item{di}{- Numeric value}

\item{With.CI}{- Logical TRUE/FALSE. Default is FALSE}
}
\value{
A graph showing the reversal if there is reversal. Or text if there is no reversal
}
\description{
############################################################################################################
 Given an 2x2 table with ai, bi, ci and di, if there are any zero cells, a continuity correction
 (CC) is added upto 0.5. This function checks if this addition of continuity correction causes
 the reversing the Odds Ratio, Relative Risk and Risk Difference.
 We provide 2 ways to add this CC. Adding CC to all the cells or only to the zero cells.
 The function will identify if there is a reversal and also give the point of reversal.
 The continuity correction can be also graphed if the option is selected.
}
\details{
The function only analyzes 2x2 tables where there is at least 1 zero cell.
By default it calculates the reversal point (if it exists) for various values of continuity correction
(CC) starting from 10^(-8) (suggested by reference [5]) all the way to 0.5. If there is a reversal due to CC
the function returns the graph showing the reversal for Relative Risk.
The function adds the CC using 2 methods.
In Method 1 - it adds CC to all the cells (called All) which is a widespread practice
In Method 2 - it adds CC to only the zero cells (called Only.Zero)
Option is provided to graph the confidence interaval (CI).
}
\examples{
## Example 1 - From reference [2] - reversal only when CC is added to the zero cells alone
ai=80; bi=1; ci=39; di=0
Plot.reversals.RR(ai,bi,ci,di, With.CI=FALSE)
##  Example 2 - From reference [3] -reversal when CC is added
ai=1; bi=33; ci=0; di=7
Plot.reversals.RR(ai,bi,ci,di, With.CI=FALSE)
## Example 3 - Above example with CI
ai=1; bi=33; ci=0; di=7
Plot.reversals.RR(ai,bi,ci,di, With.CI=TRUE)
}
\references{
[1] J Sweeting, Michael, Alexander J Sutton, and Paul C Lambert.
"What to add to nothing? Use and avoidance of continuity corrections in meta-analysis of sparse data."
Statistics in medicine 23.9 (2004): 1351-1375.
[2] Song H, Zhu J, Lu D. Molecular-targeted first-line therapy for advanced gastric cancer.
Cochrane Database of Systematic Reviews 2016, Issue 7. Art. No.: CD011461.
[3] Lensen SF, Manders M, Nastri CO, Gibreel A, Martins WP, Templer GE, Farquhar C.
Endometrial injury for pregnancy following sexual intercourse or intrauterine insemination.
Cochrane Database of Systematic Reviews 2016, Issue 6. Art. No.: CD011424.
[4] Nidorf SM, Eikelboom JW, Budgeon CA, Thompson PL.
Low-dose colchicine for secondary prevention of cardiovascular disease.
Journal of the American College of Cardiology 2013;61(4):404-10.
}
\seealso{
Other 2x2 Summary methods: \code{\link{Plot.reversals.OR}},
  \code{\link{Plot.reversals.RD}}
}
