% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/200.OS.Reversal.Association.Pattern.R
\name{Reversal.Association.Finder}
\alias{Reversal.Association.Finder}
\title{Given an input matrix, we can split it into smaller sub-matrix (min 2x2) and then find
 the Chi-squared test for each sub-matrix. The smaller matrix can "support"  or "oppose"
 (have a different conclusion at 95% confidence interaval) compared with the overall Chi-squared
 test value of the full input matrix. We count the number of times each cell supports or
 opposes the overall Chi-squared test. We also generate the possible list of sub-matrix.}
\usage{
Reversal.Association.Finder(mat)
}
\arguments{
\item{mat}{- matrix for which the sub-matrix is to be generated}
}
\value{
A dataframe with
 \item{Dimention.row}{Number of rows}
 \item{Dimention.col}{Number of columns}
 \item{Chi-sq-test.value}{Chi squared test value}
 \item{Significance at .95}{Test for significance at .95 confidence level}
}
\description{
Given an input matrix, we can split it into smaller sub-matrix (min 2x2) and then find
 the Chi-squared test for each sub-matrix. The smaller matrix can "support"  or "oppose"
 (have a different conclusion at 95% confidence interaval) compared with the overall Chi-squared
 test value of the full input matrix. We count the number of times each cell supports or
 opposes the overall Chi-squared test. We also generate the possible list of sub-matrix.
}
\details{
This can be used as an outlier detection method as well as observing the individual
cells within an IxJ table
}
\examples{
## Most influential school of Psychiatric thought and ascribed origin of schizophrenia- Agresti 1992
Eclectic=c(90,	12,	78) # Example from [reference 4]
Medical=c(13,	1,	6)
Psychoanalytic=c(19,	13,	50)
mat=rbind(Eclectic,Medical,Psychoanalytic)
Reversal.Association.Finder(mat)
}
\references{
[1] J.Berkson
Some difficulties of interpretation encountered in the application of the chi-square test
Journal of the American Statistical Association. 33,  1938, 526-536.
[2] H.W,Norton
Calculation of chi-square for complex contingency tables
Journal of the American Statistical Association. 40, 1945, 251-258.
[3] C.R. Blyth
On Simpsons paradox and the sure thing principle.
Journal of the American Statistical Association. 67, 1972, 364-366.
[4] A.Agresti
Categorical Data Analysis
(New York: Wiley & Sons 1990) pp 51-54.
}
