% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/300.NOSE-prepare-data.R
\name{Reversal.point}
\alias{Reversal.point}
\title{Given an 2x2 table with ai, bi, ci and di, if there are any zero cells, a continuity correction
 (CC) is added upto 0.5. This function checks if this addition of continuity correction causes
 the reversing the Odds Ratio, Relative Risk and Risk Difference.
 We provide 2 ways to add this CC. Adding CC to all the cells or only to the zero cells.
 The function will identify if there is a reversal and also give the point of reversal.}
\usage{
Reversal.point(ai, bi, ci, di)
}
\arguments{
\item{ai}{- Numeric value}

\item{bi}{- Numeric value}

\item{ci}{- Numeric value}

\item{di}{- Numeric value}
}
\value{
A dataframe with
 \item{Reversal.Flag}{  Text description -
 "None" if there is no reversal,
 "Both" if there is reversal when adding CC to all cells and adding CC to zero cells only,
 "Only.All" if there is reversal when adding CC to all cells only and no reversal when adding to zero cells,
 "Only.Zero" if there is reversal when adding CC to zero cells only and no reversal when adding CC to all cells }
 \item{OR.all.reversal.point}{ Value above which there will be reversal of Odds Ratio when adding CC to all the cells. If this is zero then there is no reversal}
 \item{RR.all.reversal.point }{ Value above which there will be reversal of Relative Risk when adding CC to all the cells. If this is zero then there is no reversal }
 \item{RD.all.reversal.point}{  Value above which there will be reversal of Risk Difference when adding CC to all the cells. If this is zero then there is no reversal}
 \item{OR.zero.reversal.point }{ Value above which there will be reversal of Odds Ratio when adding CC to only the zero cells. If this is zero then there is no reversal}
 \item{RR.zero.reversal.point}{ Value above which there will be reversal of Relative Risk when adding CC to only the zero cells. If this is zero then there is no reversal}
 \item{RD.zero.reversal.point }{ Value above which there will be reversal of Risk Difference when adding CC to only the zero cells. If this is zero then there is no reversal}
}
\description{
Given an 2x2 table with ai, bi, ci and di, if there are any zero cells, a continuity correction
 (CC) is added upto 0.5. This function checks if this addition of continuity correction causes
 the reversing the Odds Ratio, Relative Risk and Risk Difference.
 We provide 2 ways to add this CC. Adding CC to all the cells or only to the zero cells.
 The function will identify if there is a reversal and also give the point of reversal.
}
\details{
The function only analyzes 2x2 tables where there is at least 1 zero cell.
By default it calculates the reversal point (if it exists) for various values of continuity correction
(CC) starting from 10^(-8) (suggested by reference [5]) all the way to 0.5. If there is a reversal due to CC
the function returns the value above which there will be a reversal for OR, RR and RD.
The function adds the CC using 2 methods.
In Method 1 - it adds CC to all the cells (called All) which is a widespread practice
In Method 2 - it adds CC to only the zero cells (called Only.Zero)
If the rerutn values with the dataframe are zero it means that there was no reversal for CC upto 0.5.
}
\examples{
## Example 1 - From reference [2] - reversal only when CC is added to the zero cells alone
ai=80; bi=1; ci=39; di=0
Reversal.point(ai,bi,ci,di)
##  Example 2 - From reference [3] -reversal when CC is added
ai=1; bi=33; ci=0; di=7
Reversal.point(ai,bi,ci,di)
##  Example 3 - From reference [4] - no reversal when CC is added
ai=0; bi=282; ci=1; di=249
Reversal.point(ai,bi,ci,di)
}
\references{
[1] J Sweeting, Michael, Alexander J Sutton, and Paul C Lambert.
"What to add to nothing? Use and avoidance of continuity corrections in meta-analysis of sparse data."
Statistics in medicine 23.9 (2004): 1351-1375.
[2] Song H, Zhu J, Lu D. Molecular-targeted first-line therapy for advanced gastric cancer.
Cochrane Database of Systematic Reviews 2016, Issue 7. Art. No.: CD011461.
[3] Lensen SF, Manders M, Nastri CO, Gibreel A, Martins WP, Templer GE, Farquhar C.
Endometrial injury for pregnancy following sexual intercourse or intrauterine insemination.
Cochrane Database of Systematic Reviews 2016, Issue 6. Art. No.: CD011424.
[4] Nidorf SM, Eikelboom JW, Budgeon CA, Thompson PL.
Low-dose colchicine for secondary prevention of cardiovascular disease.
Journal of the American College of Cardiology 2013;61(4):404-10.
}
\seealso{
\code{\link[metafor]{rma}}  provides OR, RR and RD values for various inputs, with some control
of the cc via the "to=" option. However, we suggest using metafor by passing the continuity corrected values
to replicate the results of this function, since adding only to zero cell is not an option.

Other IxJ Inference methods: \code{\link{Local.Odds.Ratio}},
  \code{\link{Nominal.Odds.Ratio}},
  \code{\link{Plot.heatmap}},
  \code{\link{Subtable.Odds.Ratio}},
  \code{\link{generate.heatmap.matrix}}
}
