% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/602.Subtable-odds-ratio.R
\name{Subtable.Odds.Ratio}
\alias{Subtable.Odds.Ratio}
\title{Given an input matrix, we can split it into smaller 2x2 sub-matrix and then find
 the Odds Ratio  for each sub-matrix (subtable).}
\usage{
Subtable.Odds.Ratio(mat)
}
\arguments{
\item{mat}{- matrix for which the sub-matrix is to be generated.}
}
\value{
A of dataframes with
 \item{Odds.Ratio}{  Odds Ratio of the 2x2 subtable}
 \item{a }{ The value of a taken from the input table}
 \item{b }{ The value of b taken from the input table }
 \item{c }{ The value of c taken from the input table }
 \item{d }{ The value of d taken from the input table }
 \item{Index1RC }{ Index reference from the input matrix for a (Row,Column) }
 \item{Index2RC }{ Index reference from the input matrix for b (Row,Column) }
 \item{Index3RC }{ Index reference from the input matrix for c (Row,Column) }
 \item{Index4RC }{ Index reference from the input matrix for d (Row,Column) }
}
\description{
Given an input matrix, we can split it into smaller 2x2 sub-matrix and then find
 the Odds Ratio  for each sub-matrix (subtable).
}
\details{
The function calculates the odds ratio for every 2x2 sub-matrix for the input matrix.
This can be used as input to models. If any of the cells a,b,c or d
is zero, a continuity correction of 0.5 is added to all the cells and the odds ratio is calculated.
}
\examples{
##  Example data from [reference 1]
Drills=c(2,	10,	4,	2 )
Pots= c(3,	8,	4,	6)
Grinding.Stones=c( 13, 5, 3, 9)
Point.Fragments=c(20, 36, 19, 20)
mat=rbind(Drills,Pots,Grinding.Stones,Point.Fragments)
Subtable.Odds.Ratio(mat)
}
\references{
[1] Mosteller F, Parunak A (2006)
Identifying extreme cells in a sizable contingency table: Probabilistic and exploratory approaches.
In: Hoaglin DC, Mosteller F, Tukey JW (eds) Exploring Data Tables, Trends, and Shapes,
John Wiley & Sons, pp 189-224
}
\seealso{
Other IxJ Inference methods: \code{\link{Local.Odds.Ratio}},
  \code{\link{Nominal.Odds.Ratio}},
  \code{\link{Plot.heatmap}}, \code{\link{Reversal.point}},
  \code{\link{generate.heatmap.matrix}}
}
