% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/500.SummaryMeasuresCategoricalData20170518.R
\name{SummaryData}
\alias{SummaryData}
\title{The summary of measures for categorical data}
\usage{
SummaryData(inpmat, min = NULL, max = NULL)
}
\arguments{
\item{inpmat}{- The input matrix}

\item{min}{- The minimum of the range of interest}

\item{max}{- The minimum of the range of interest}
}
\value{
A list of dataframes
 \item{SumZeroData}{ Count of rows and columns that contain all zeros, in a given matrix}
 \item{SumCount}{ Count of cells in a given matrix with value of zero or value within the range specified as min and max}
 \item{SumMinMaxRange}{  Calculates the min, max and range of values in a given matrix}
 \item{SumMedianSD}{  Calculates median, SD, upper limit, lower limit and cell values within these limits}
}
\description{
The summary of measures for categorical data
}
\details{
The function calculates the summary using
}
\examples{
x=c(7,61,55,129,489,570,475,431,293,154,38,12)
inpmat = cbind(x[1:4],x[5:8],x[9:12])
min = 0
max = 150
SummaryData(inpmat, min, max)
SummaryData(inpmat)
}
\seealso{
Other The summary of measures for categorical data: \code{\link{SumCount}},
  \code{\link{SumMedianSD}}, \code{\link{SumMinMaxRange}},
  \code{\link{SumZeroData}}
}
\author{
Subbiah and Balakrishna  S Kesavan
}
