% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/500.SummaryMeasuresCategoricalData20170518.R
\name{SumMinMaxRange}
\alias{SumMinMaxRange}
\title{Calculates the min, max and range of values in a given matrix}
\usage{
SumMinMaxRange(inpmat)
}
\arguments{
\item{inpmat}{- The input matrix}
}
\value{
A dataframe that provides the min, max and range of values in a given matrix
 \item{Feature}{ Feature description -
 " Max value in table" showing the maximum of the input cells
 "Min value in table" showing the minimum of the input cells
 "Range of values in table" showing the difference between the maximum and minimum"}
 \item{Value}{ Calculated value of features}
}
\description{
Calculates the min, max and range of values in a given matrix
}
\details{
The function calculates the min, max and range of values in a given matrix
}
\examples{
x=c(7,61,55,129,489,570,475,431,293,154,38,12)
inpmat = cbind(x[1:4],x[5:8],x[9:12])
SumMinMaxRange(inpmat)
}
\seealso{
Other The summary of measures for categorical data: \code{\link{SumCount}},
  \code{\link{SumMedianSD}}, \code{\link{SumZeroData}},
  \code{\link{SummaryData}}
}
