% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/500.SummaryMeasuresCategoricalData20170518.R
\name{SumZeroData}
\alias{SumZeroData}
\title{Calculates the count of rows and columns that contain all zeros, in a given matrix}
\usage{
SumZeroData(inpmat, start.col = NULL, end.col = NULL)
}
\arguments{
\item{inpmat}{- The input matrix}

\item{start.col}{- The minimum of the range of interest. Starting column number}

\item{end.col}{- The minimum of the range of interest. Ending column number}
}
\value{
A dataframe that provides a count of rows and columns that contain all
zeros, in a given matrix
 \item{Feature}{Feature description -
 '# of all-zero rows' showing the count of rows with all zeros
 '# of all-zero columns' showing the count of columns with all zeros}
 \item{Value}{ Calculated value of counts}
}
\description{
Calculates the count of rows and columns that contain all zeros, in a given matrix
}
\details{
The function calculates the number of rows and columns which are all zeros, in the given range
}
\examples{
x=c(7,61,55,129,489,570,475,431,293,154,38,12)
inpmat = cbind(x[1:4],x[5:8],x[9:12])
SumZeroData(inpmat)
###Example with some zeros
inpmat = matrix(c(0, 0, 0, 1, 0, 7), nrow=3, ncol=2)
SumZeroData(inpmat)
}
\seealso{
Other The summary of measures for categorical data: \code{\link{SumCount}},
  \code{\link{SumMedianSD}}, \code{\link{SumMinMaxRange}},
  \code{\link{SummaryData}}
}
\author{
Subbiah and Balakrishna  S Kesavan
}
